Ok now i think it works (not sure but there's is hope)

+++ src/magic.h
@@ -25,8 +25,8 @@ public:
     std::array<Bitboard, SQUARE_NB> bishop_masks;
     std::array<Bitboard, SQUARE_NB> rook_magics;
     std::array<Bitboard, SQUARE_NB> bishop_magics;
-    std::array<uint8_t, SQUARE_NB> rook_shifts;
-    std::array<uint8_t, SQUARE_NB> bishop_shifts;
+    std::array<int, SQUARE_NB> rook_shifts;
+    std::array<int, SQUARE_NB> bishop_shifts;
     std::array<std::vector<Bitboard>, SQUARE_NB> rook_move_table;
     std::array<std::vector<Bitboard>, SQUARE_NB> bishop_move_table;
 };
